BF: Fix sound component stop value handling without '$'
Addressed an issue in the sound component where code expressions for the
stop value without preceding '$' were causing errors.

This change also moves `canBeNumeric()` from `experiment.componenets._base` to
`experiment.utils`.
