+++ psychopy/experiment/components/sound/__init__.py
@@ -10,6 +10,7 @@ Distributed under the terms of the GNU General Public License (GPL).
 from pathlib import Path
 from psychopy.experiment.components import BaseDeviceComponent, Param, getInitVals, \
     _translate
+from psychopy.experiment.utils import canBeNumeric
 from psychopy.tools.audiotools import knownNoteNames
 
 
@@ -139,7 +140,7 @@ class SoundComponent(BaseDeviceComponent):
     def writeInitCode(self, buff):
         # replaces variable params with sensible defaults
         inits = getInitVals(self.params)
-        if '$' in str(inits['stopVal'].val):
+        if not canBeNumeric(inits['stopVal'].val):
             inits['stopVal'].val = -1
         else:
             if inits['stopVal'].val in ['', None, 'None']:
@@ -177,7 +178,7 @@ class SoundComponent(BaseDeviceComponent):
     def writeInitCodeJS(self, buff):
         # replaces variable params with sensible defaults
         inits = getInitVals(self.params)
-        if '$' in inits['stopVal'].val:
+        if not canBeNumeric(inits['stopVal'].val):
             inits['stopVal'].val = -1
         elif inits['stopVal'].val in ['', None, 'None']:
             inits['stopVal'].val = -1
